Split honeycomb to api and core
