+++ honeycomb/api/src/main/java/org/commonjava/o11yphant/honeycomb/SpanContext.java
