+++ honeycomb/core/src/main/java/org/commonjava/o11yphant/honeycomb/RootSpanFields.java
