+++ honeycomb/core/src/main/java/org/commonjava/o11yphant/honeycomb/impl/CassandraConnectionRootSpanFields.java
