+++ honeycomb/core/src/test/java/org/commonjava/o11yphant/honeycomb/impl/JVMRootSpanFieldsTest.java
