+++ metrics/core/src/main/java/org/commonjava/o11yphant/metrics/RequestContextHelper.java
@@ -17,11 +17,8 @@ package org.commonjava.o11yphant.metrics;
 
 import org.commonjava.cdi.util.weft.ThreadContext;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.RetentionPolicy.SOURCE;
+import static org.commonjava.o11yphant.metrics.RequestContextConstants.END_NANOS;
+import static org.commonjava.o11yphant.metrics.RequestContextConstants.RAW_IO_WRITE_NANOS;
 
 /**
  * The scope annotations (Thread, Header, MDC) tell where the constant is available/used. The static methods are used
@@ -29,8 +26,6 @@ import static java.lang.annotation.RetentionPolicy.SOURCE;
  */
 public class RequestContextHelper
 {
-    public static final String HEADER_COMPONENT_ID = "component-id";
-
     public static void setContext( final String key, final Object value )
     {
         org.slf4j.MDC.put( key, String.valueOf( value ) );
@@ -74,113 +69,5 @@ public class RequestContextHelper
     {
         return getContext( RAW_IO_WRITE_NANOS, 0L );
     }
-
-    // Scope annotations
-
-    @Target( { FIELD } )
-    @Retention( SOURCE )
-    private @interface Header{}
-
-    @Target( { FIELD } )
-    @Retention( SOURCE )
-    private @interface MDC{}
-
-    @Target( { FIELD } )
-    @Retention( SOURCE )
-    private @interface Thread{}
-
-    //
-
-    @Thread
-    public static final String RAW_IO_WRITE_NANOS = "raw-io-write-nanos";
-
-    @Thread
-    public static final String RAW_IO_READ_NANOS = "raw-io-read-nanos";
-
-    @Thread
-    public static final String END_NANOS = "latency-end-nanos";
-
-    @Thread @MDC
-    public static final String REST_CLASS_PATH = "REST-class-path";
-
-    @Thread @MDC
-    public static final String REST_METHOD_PATH = "REST-method-path";
-
-    @Thread @MDC
-    public static final String REST_ENDPOINT_PATH = "REST-endpoint-path";
-
-    @Thread @MDC
-    public static final String REST_CLASS = "REST-class";
-
-    @MDC
-    public static final String CONTENT_TRACKING_ID = "tracking-id";
-
-    @MDC
-    public static final String ACCESS_CHANNEL = "access-channel";
-
-    @MDC
-    public static final String REQUEST_LATENCY_NS = "request-latency-ns";
-
-    @Thread
-    public static final String REQUEST_LATENCY_MILLIS = "latency_ms";
-
-    @MDC
-    public static final String REQUEST_PHASE = "request-phase";
-
-    @Thread @MDC
-    public static final String PACKAGE_TYPE = "package-type";
-
-    @Thread @MDC
-    public static final String METADATA_CONTENT = "metadata-content";
-
-    @Thread @MDC
-    public static final String CONTENT_ENTRY_POINT = "content-entry-point";
-
-    @Thread @MDC
-    public static final String HTTP_METHOD = "http-method";
-
-    @Thread @MDC
-    public static final String HTTP_REQUEST_URI = "http-request-uri";
-
-    @Thread @MDC
-    public static final String PATH = "path";
-
-    @Thread @MDC
-    public static final String HTTP_STATUS = "http-status";
-
-    @Header @MDC
-    public static final String COMPONENT_ID = HEADER_COMPONENT_ID; // "component-id";
-
-    @Header
-    public static final String X_FORWARDED_FOR = "x-forwarded-for";
-
-    @Header @Thread @MDC
-    public static final String EXTERNAL_TRACE_ID = "external-id";
-
-    @Header
-    public static final String SPAN_ID_HEADER = "span-id";
-
-    @Thread @MDC
-    public static final String REQUEST_PARENT_SPAN = "parent-span";
-
-    @Thread @MDC
-    public static final String CLIENT_ADDR = "client-addr";
-
-    @Thread @MDC
-    public static final String INTERNAL_ID = "internal-id";
-
-    @Thread @MDC
-    public static final String TRACE_ID = "trace-id";
-
-    @Thread
-    public static final String FORCE_METERED = "force-metered";
-
-    @Thread
-    public static final String IS_METERED = "is-metered";
-
-    // these are well-known values we'll be using in our log aggregation filters
-    public static final String REQUEST_PHASE_START = "start";
-
-    public static final String REQUEST_PHASE_END = "end";
 
 }
