+++ pom.xml
@@ -81,6 +81,17 @@
 
             <dependency>
                 <groupId>org.commonjava.util</groupId>
+                <artifactId>o11yphant-honeycomb-api</artifactId>
+                <version>1.3-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>org.commonjava.util</groupId>
+                <artifactId>o11yphant-honeycomb-core</artifactId>
+                <version>1.3-SNAPSHOT</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.commonjava.util</groupId>
                 <artifactId>o11yphant-metrics</artifactId>
                 <version>1.3-SNAPSHOT</version>
             </dependency>
